project/add_pictureTag/pyproject.toml
[project]
name = "add-picturetag"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "opencv-python>=4.11.0.86",
    "pillow>=11.2.1",
]
